Drop original STS
