+++ src/online/sts_config.h.in
